Move legacy support classes to another jar
