+++ launcher/minecraft/MinecraftInstance.h
@@ -132,6 +132,7 @@ public:
     /// get arguments passed to java
     QStringList javaArguments();
     QString getLauncher();
+    bool shouldApplyOnlineFixes();
 
     /// get variables for launch command variable substitution/environment
     QMap<QString, QString> getVariables() override;
