+++ libraries/launcher/CMakeLists.txt
@@ -11,25 +11,33 @@ set(SRC
     org/prismlauncher/launcher/Launcher.java
     org/prismlauncher/launcher/impl/AbstractLauncher.java
     org/prismlauncher/launcher/impl/StandardLauncher.java
-    org/prismlauncher/launcher/impl/legacy/LegacyLauncher.java
-    org/prismlauncher/launcher/impl/legacy/LegacyFrame.java
     org/prismlauncher/exception/ParameterNotFoundException.java
     org/prismlauncher/exception/ParseException.java
-    org/prismlauncher/fix/Fixes.java
-    org/prismlauncher/fix/online/OnlineFixes.java
-    org/prismlauncher/fix/online/SkinFix.java
-    org/prismlauncher/fix/online/Handler.java
-    org/prismlauncher/utils/Base64.java
-    org/prismlauncher/utils/JsonParser.java
     org/prismlauncher/utils/Parameters.java
     org/prismlauncher/utils/ReflectionUtils.java
-    org/prismlauncher/utils/api/MojangApi.java
-    org/prismlauncher/utils/api/Texture.java
     org/prismlauncher/utils/logging/Level.java
     org/prismlauncher/utils/logging/Log.java
-    org/prismlauncher/utils/url/UrlUtils.java
-    org/prismlauncher/utils/url/CustomUrlConnection.java
-    net/minecraft/Launcher.java
+    org/prismlauncher/legacy/LegacyProxy.java
 )
+
+set(LEGACY_SRC
+    legacy/org/prismlauncher/legacy/LegacyFrame.java
+    legacy/org/prismlauncher/legacy/LegacyLauncher.java
+    legacy/org/prismlauncher/legacy/fix/online/Handler.java
+    legacy/org/prismlauncher/legacy/fix/online/OnlineFixes.java
+    legacy/org/prismlauncher/legacy/fix/online/SkinFix.java
+    legacy/org/prismlauncher/legacy/utils/Base64.java
+    legacy/org/prismlauncher/legacy/utils/JsonParseException.java
+    legacy/org/prismlauncher/legacy/utils/JsonParser.java
+    legacy/org/prismlauncher/legacy/utils/api/MojangApi.java
+    legacy/org/prismlauncher/legacy/utils/api/Texture.java
+    legacy/org/prismlauncher/legacy/utils/url/CustomUrlConnection.java
+    legacy/org/prismlauncher/legacy/utils/url/UrlUtils.java
+    legacy/net/minecraft/Launcher.java
+    legacy/org/prismlauncher/legacy/LegacyProxy.java
+)
+
 add_jar(NewLaunch ${SRC})
+add_jar(NewLaunchLegacy ${LEGACY_SRC} INCLUDE_JARS NewLaunch)
 install_jar(NewLaunch "${JARS_DEST_DIR}")
+install_jar(NewLaunchLegacy "${JARS_DEST_DIR}")
